pyproject.toml
[build-system]
requires = ["hatchling >= 1.21.0"]
build-backend = "hatchling.build"

[project]
name = "yak-server"
description = "Football bet rest/graphql server"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [{name = "Guillaume Le Pape", email = "gui.lepape25@gmail.com"}]
keywords = ["mysql", "api", "rest", "graphql"]
classifiers = [
    "Operating System :: Unix",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Framework :: Pydantic",
    "Framework :: Pydantic :: 2",
    "License :: OSI Approved :: MIT License",
    "Topic :: Database :: Database Engines/Servers",
    "Topic :: Database",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "PyJWT ==2.9.0",
    "pymysql[rsa] ==1.1.1",
    "argon2-cffi ==23.1.0",
    "fastapi ==0.115.4",
    "typer ==0.12.5",
    "SQLAlchemy ==2.0.36",
    "SQLAlchemy-Utils ==0.41.2",
    "pendulum ==3.0.0",
    "pydantic-settings ==2.6.1",
    "strawberry-graphql ==0.247.1",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/yak-toto/yak-server"
Repository = "https://github.com/yak-toto/yak-server"

[project.optional-dependencies]
db_migration = ["alembic ==1.13.3"]
profiling = ["yappi ==1.6.0"]
sync = ["beautifulsoup4[lxml] ==4.12.3", "httpx ==0.27.2"]

[project.scripts]
yak = "yak_server.cli:app"

[tool.hatch.version]
path = "yak_server/__init__.py"

[tool.hatch.build.targets.wheel.hooks.custom]
path = "scripts/build_hooks.py"

[tool.coverage.run]
branch = true

[tool.coverage.report]
omit = ["yak_server/database/migrations/*", "scripts/build_hooks.py"]
exclude_lines = ["if TYPE_CHECKING:", "pragma: no cover"]

[tool.ruff]
line-length = 100
target-version = "py39"
fix = true
preview = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",
    "ANN101",  # missing annotation for self?
    "ANN002",
    "ANN003",
    "ANN102",
    "ANN202",
    # shadowing builtins
    "A002",
    "A003",
    "PLR",
    "G004",
    "FA100",
    "COM812",
    "ISC001",
    "CPY001",
]

[tool.ruff.lint.mccabe]
max-complexity = 11

[tool.ruff.lint.per-file-ignores]
"yak_server/cli/*" = ["T20"]
"yak_server/database/migrations/*" = ["ANN001", "ANN201"]
"yak_server/helpers/errors.py" = ["S105"]
"tests/*" = ["S101"]
"testing/*" = ["S101"]
"scripts/*" = ["T201"]

[tool.mypy]
strict = true
pretty = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin", "strawberry.ext.mypy_plugin"]

[[tool.mypy.overrides]]
module = "sqlalchemy_utils"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yappi"
ignore_missing_imports = true

[tool.pytest.ini_options]
log_file = "yak_test.log"
# log_file_format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
log_file_level = "INFO"
env = [
    "MYSQL_DB=yak_toto_test",
    "PROFILING=0",
]

[tool.importlinter]
root_package = "yak_server"

[[tool.importlinter.contracts]]
name = "yak_server.database should not import other yak_server modules"
type = "independence"
modules = ["yak_server.database"]

[[tool.importlinter.contracts]]
name = "yak_server.helpers should not import yak_server.{v1,v2,cli}"
type = "forbidden"
source_modules = ["yak_server.helpers"]
forbidden_modules = ["yak_server.v1", "yak_server.v2", "yak_server.cli"]

[[tool.importlinter.contracts]]
name = "yak_server.v1.helpers should not import yak_server.v1.models and yak_server.v1.routers"
type = "forbidden"
source_modules = ["yak_server.v1.helpers"]
forbidden_modules = ["yak_server.v1.models", "yak_server.v1.routers"]

[[tool.importlinter.contracts]]
name = "yak_server.v1 and yak_server.v2 should be independent"
type = "independence"
modules = ["yak_server.v1", "yak_server.v2"]
